requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
httpx>=0.27.0
python-multipart>=0.0.7
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langgraph>=0.1.0
torch>=2.2.0
transformers>=4.40.0
docling>=2.0.0
pinecone
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
playwright>=1.41.0
structlog>=24.1.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
pyyaml>=6.0.0
firecrawl-py>=1.0.0
